+++ engine/src/flutter/impeller/renderer/backend/gles/capabilities_gles.h
@@ -117,6 +117,9 @@ class CapabilitiesGLES final
   bool SupportsPrimitiveRestart() const override;
 
   // |Capabilities|
+  bool Supports32BitPrimitiveIndices() const override;
+
+  // |Capabilities|
   bool SupportsExtendedRangeFormats() const override;
 
   // |Capabilities|
@@ -146,6 +149,7 @@ class CapabilitiesGLES final
   bool supports_decal_sampler_address_mode_ = false;
   bool supports_offscreen_msaa_ = false;
   bool supports_implicit_msaa_ = false;
+  bool supports_32bit_primitive_indices_ = false;
   bool is_angle_ = false;
   bool is_es_ = false;
   PixelFormat default_glyph_atlas_format_ = PixelFormat::kUnknown;