+++ engine/src/flutter/impeller/renderer/backend/vulkan/capabilities_vk.h
@@ -261,6 +261,9 @@ class CapabilitiesVK final : public Capabilities,
   bool SupportsPrimitiveRestart() const override;
 
   // |Capabilities|
+  bool Supports32BitPrimitiveIndices() const override;
+
+  // |Capabilities|
   bool SupportsExtendedRangeFormats() const override;
 
   // |Capabilities|