+++ engine/src/flutter/shell/platform/android/test/io/flutter/embedding/engine/image/ImageDecoderHeifPre36ImplTest.java
@@ -8,7 +8,7 @@ import static org.junit.Assert.assertEquals;
 
 import android.graphics.Bitmap;
 import android.graphics.Color;
-import android.media.ExifInterface;
+import androidx.exifinterface.media.ExifInterface;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import io.flutter.Build;
 import java.nio.ByteBuffer;