+++ engine/src/flutter/shell/platform/android/test/io/flutter/view/AccessibilityBridgeTest.java
@@ -2513,6 +2513,7 @@ public class AccessibilityBridgeTest {
     int platformViewId = -1;
     int scrollChildren = 0;
     int scrollIndex = 0;
+    int traversalParent = -1;
     float scrollPosition = 0.0f;
     float scrollExtentMax = 0.0f;
     float scrollExtentMin = 0.0f;
@@ -2543,6 +2544,14 @@ public class AccessibilityBridgeTest {
           0.0f, 0.0f, 1.0f, 0.0f,
           0.0f, 0.0f, 0.0f, 1.0f
         };
+    float[] hitTestTransform =
+        new float[] {
+          1.0f, 0.0f, 0.0f, 0.0f,
+          0.0f, 1.0f, 0.0f, 0.0f,
+          0.0f, 0.0f, 1.0f, 0.0f,
+          0.0f, 0.0f, 0.0f, 1.0f
+        };
+
     final List<TestSemanticsNode> children = new ArrayList<TestSemanticsNode>();
 
     public void addChild(TestSemanticsNode child) {
@@ -2574,6 +2583,7 @@ public class AccessibilityBridgeTest {
       bytes.putInt(platformViewId);
       bytes.putInt(scrollChildren);
       bytes.putInt(scrollIndex);
+      bytes.putInt(traversalParent);
       bytes.putFloat(scrollPosition);
       bytes.putFloat(scrollExtentMax);
       bytes.putFloat(scrollExtentMin);
@@ -2616,12 +2626,17 @@ public class AccessibilityBridgeTest {
       for (int i = 0; i < 16; i++) {
         bytes.putFloat(transform[i]);
       }
+      // hitTestTransform.
+      for (int i = 0; i < 16; i++) {
+        bytes.putFloat(hitTestTransform[i]);
+      }
       // children in traversal order.
       bytes.putInt(children.size());
       for (TestSemanticsNode node : children) {
         bytes.putInt(node.id);
       }
       // children in hit test order.
+      bytes.putInt(children.size());
       for (TestSemanticsNode node : children) {
         bytes.putInt(node.id);
       }