+++ packages/flutter/lib/src/rendering/custom_paint.dart
@@ -696,17 +696,15 @@ class RenderCustomPaint extends RenderProxyBox {
       return true;
     }());
 
-    final List<CustomPainterSemantics> backgroundSemantics = _backgroundSemanticsBuilder != null
-        ? _backgroundSemanticsBuilder!(size)
-        : const <CustomPainterSemantics>[];
+    final List<CustomPainterSemantics> backgroundSemantics =
+        _backgroundSemanticsBuilder?.call(size) ?? const <CustomPainterSemantics>[];
     _backgroundSemanticsNodes = _updateSemanticsChildren(
       _backgroundSemanticsNodes,
       backgroundSemantics,
     );
 
-    final List<CustomPainterSemantics> foregroundSemantics = _foregroundSemanticsBuilder != null
-        ? _foregroundSemanticsBuilder!(size)
-        : const <CustomPainterSemantics>[];
+    final List<CustomPainterSemantics> foregroundSemantics =
+        _foregroundSemanticsBuilder?.call(size) ?? const <CustomPainterSemantics>[];
     _foregroundSemanticsNodes = _updateSemanticsChildren(
       _foregroundSemanticsNodes,
       foregroundSemantics,
@@ -1021,6 +1019,12 @@ class RenderCustomPaint extends RenderProxyBox {
     if (properties.identifier != null) {
       config.identifier = properties.identifier!;
     }
+    if (properties.traversalParentIdentifier != null) {
+      config.traversalParentIdentifier = properties.traversalParentIdentifier;
+    }
+    if (properties.traversalChildIdentifier != null) {
+      config.traversalChildIdentifier = properties.traversalChildIdentifier;
+    }
     if (properties.tooltip != null) {
       config.tooltip = properties.tooltip!;
     }
@@ -1041,6 +1045,9 @@ class RenderCustomPaint extends RenderProxyBox {
     }
     if (config.validationResult != properties.validationResult) {
       config.validationResult = properties.validationResult;
+    }
+    if (properties.hitTestBehavior != null) {
+      config.hitTestBehavior = properties.hitTestBehavior!;
     }
     if (properties.inputType != null) {
       config.inputType = properties.inputType!;